requirements.txt
google-generativeai>=0.3.2
openai>=1.12.0
aiohttp>=3.9.3
numpy>=1.24.3
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
pyyaml>=6.0.1
requests>=2.31.0
tenacity>=8.2.3
tiktoken>=0.5.2
typing-extensions>=4.9.0
urllib3>=2.2.0
validators>=0.22.0